feat: adicionar novos campos e métodos ao modelo de usuário, incluindo CPF, endereço, telefone e timestamps de criação e atualização
