--- api/model/__init__.py
@@ -1,12 +1,0 @@
-from .user_model import User
-from .client_model import Client
-from .vehicle_model import Vehicle
-from .rent_model import Rental
-
-
-__All__ = [
-    "User",
-    "Client",
-    "Vehicle",
-    "Rent",
-]
