--- api/serializers/__init__.py
@@ -1,10 +1,0 @@
-from .client_serializer import ClientDetailsSerializer
-from .user_serializer import UserSerializer, UserUpdateSerializer
-from .authentication_serializer import MyTokenObtainPairSerializer
-
-__All__ = [
-    "ClientSerializer",
-    "UserSerializer",
-    "MyTokenObtainPairSerializer",
-    "UserUpdateSerializer"
-]
