--- api/serializers/authentication_serializer.py
@@ -1,13 +1,0 @@
-from rest_framework import serializers
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-
-class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
-    @classmethod
-    def get_token(cls, user):
-        token = super().get_token(user)
-        
-        token['username'] = user.username
-        token['email'] = user.email
-        token['name'] = user.name
-        
-        return token
