--- api/serializers/user_serializer.py
@@ -1,45 +1,0 @@
-from rest_framework import serializers
-
-from api.model.user_model import User
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('id', 'username', 'name', 'email', 'cpf', 'address', 'phone', 'created_at', 'updated_at')
-        extra_kwargs = {
-            'password': {'write_only': True},
-            'created_at': {'read_only': True},
-            'updated_at': {'read_only': True}
-            }
-
-    def create(self, validated_data):
-        user = User.objects.create_user(**validated_data)
-        return user
-    
-
-class UserUpdateSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('id', 'username', 'name', 'email', 'cpf', 'address', 'phone', 'created_at', 'updated_at')
-        extra_kwargs = {
-            'password': {'write_only': True},
-            'created_at': {'read_only': True},
-            'updated_at': {'read_only': True}
-            }
-
-    def update(self, instance, validated_data):
-        instance.username = validated_data.get('username', instance.username)
-        instance.name = validated_data.get('name', instance.name)
-        instance.email = validated_data.get('email', instance.email)
-        instance.cpf = validated_data.get('cpf', instance.cpf)
-        instance.address = validated_data.get('address', instance.address)
-        instance.phone = validated_data.get('phone', instance.phone)
-        instance.save()
-        return instance
-
-
-class UserListSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('id', 'username', 'name', 'email', 'cpf', 'address', 'phone', 'created_at', 'updated_at')
