--- api/views/__init__.py
@@ -1,25 +1,0 @@
-from .views import ClientCreateView, UserCreateView, VehicleCreateView, RentCreateView, UserListView, ClientListView, ClientDetailView, ClientWithUserView, RentListView, UserUpdateView, RentDeleteView, VehicleDeleteView, VehicleListView, VehicleListByCarView, VehicleListByMotoView, RentServiceUpdateView,ClientDeleteView,UserDeleteView
-from .authentication_view import LoginView, LogoutView
-
-__All__ = [
-    "ClientCreateView",
-    "UserCreateView",
-    "VehicleCreateView",
-    "RentCreateView",
-    "UserListView",
-    "ClientListView",
-    "ClientDetailView",
-    "ClientWithUserView",
-    "RentListView",
-    "UserUpdateView",
-    "RentDeleteView",
-    "VehicleDeleteView",
-    "VehicleListView",
-    "VehicleListByCarView",
-    "VehicleListByMotoView",
-    "RentServiceUpdateView",
-    "ClientDeleteView",
-    "UserDeleteView",
-    "LoginView",
-    "LogoutView",
-]
